+++ banyand/measure/metadata.go
@@ -265,8 +265,8 @@ type supplier struct {
 	pipeline queue.Queue
 	omr      observability.MetricsRegistry
 	l        *logger.Logger
-	option   option
 	path     string
+	option   option
 }
 
 func newSupplier(path string, svc *service) *supplier {
@@ -311,6 +311,7 @@ func (s *supplier) OpenDB(groupSchema *commonv1.Group) (resourceSchema.DB, error
 		TTL:                            storage.MustToIntervalRule(groupSchema.ResourceOpts.Ttl),
 		Option:                         s.option,
 		SeriesIndexFlushTimeoutSeconds: s.option.flushTimeout.Nanoseconds() / int64(time.Second),
+		SeriesIndexCacheMaxBytes:       int(s.option.seriesCacheMaxSize),
 		StorageMetricsFactory:          factory,
 	}
 	return storage.OpenTSDB(